more 3D modules added
added some 3D modules …

+++ python/jtlib/modules/Jt_Modules_Saadia/label_mask_3d.py
@@ -56,8 +56,7 @@ def label_mask_3d(mask, plot=False):
     
             output1.append(label_image) 
             
-        output = np.int32(np.dstack(output1))
-        
+        output = np.dstack(output1)  
         # print output.shape
         
         output = {'label_image': output}
